extract OracleTokenScan into a separate module;
